Refactor : [WUMO-316] Image 서비스 책임 변경
- Service -> Repository 로 책임 변경
- 추후 다른 서비스에서 이미지 삭제 요청을 하기 위해 Repo 사용
